Rename project to superimpose-hand
The project name has been changed to movefinger to superimpose-hand to better describe its tasks.
The movefinger project is the basis of the current new project and it's not considered closed.
Another repository will be started as soon as possible containing all the basic iCub tasks like moving a hand to a specific position, opening or closing the hand, etc.
The new repository name will be named move-view-hand and will be considered as a template project to use for developing iCub-related project based on moving upper body iCub limbs, using the cameras and gathering associated data (pose, images, etc.) for offline simulations and correctness tests.