Upgrade to new version with OpenGL support
First commit toward a new release of the module.

The model now uses OpenGL capabilities to render a 3D CAD model of the iCub palm and to superimpose it onto the camera images.
The current version only supports the right hand palm of the iCub robot seen from the left eye only.
The code is still not optimized and needs to be refactored for a complete (and smooht) integration of OpenGL API within a YARP-structured module.