Integrate thrift IDL to handle RPC services
Thrift IDL has been fully integrated with the module in order
to improve the user experience while using the RPC.
In particular, IDL has been used to auto-generate code for
RPC services and associated documentation.
As a result, the user experience has much improved, as well as
code readability and maintainability on the service side.