Add rendering options to hand superimposition
IDL is used to provide dedicated services to the rendering thread
in order to provide real-time rendering options.
In particular, the user can now set whether it prefers to render
the hand as a wireframe or as a filled object, to view or hide
the background (i.e. the camera images) and the mipmaps filtering
technique. As a result, it is now possible to have different
rendering results and a flexible way to implement new rendering
features through IDL services.