Refactor code to separate h/cpp files
Most of the code contained in main.cpp has been dived in separe
header and cpp files. This is a first step for dividing the actual code
in a standalone program for YARP/iCub and a general AR/VR framework.
Furthermore, the CMakeLists.txt has been refactored to be more clear and
clean.