Fix resource deallocation order in SuperimposeHand
The interruptModule() call in SuperimposeHand has been removed.
Previously, interruptModule() was in charge of stopping both
SuperimposeHandCADThread and SuperimposeHandSkeletonThread if they were
running. The problem is in that this method is asynchronous w.r.t.
the close function which is invoked immediatly after updateModule()
returns. This possibly cause the threads to be deleated by the close() method
of RFModule while still stopping in interruptModule() call.
interruptModule() has been removed from SuperimposeHand and slightly
modified in SuperimposeHandCADThread and SuperimposeHandSkeletonThread
as well.