Integrate YARP RFModule threaded functionalities
The new YARP RFModule includes thread-ification capabilities
so that the tasks of an RFModule can be run on a separate thread
without blocking the main thread.
This is of particular advantage when using thread-safe GLFW
(or of similar libraries) function calls for using OpenGL.
As a result, SuperimposeHandThread.h and SuperimposeHandThread.cpp
have been deleted.