Generate new thread classes deprecating old ones
The following files are now considered (to-be) deprecated:
 - SuperimposeiCubHand.h
 - SiCHCAD.h
 - SiCHSkeleton.h
 - SuperimposeiCubHand.cpp
 - SiCHCAD.cpp
 - SiCHSkeleton.cpp
These files are still in their original include/src directories
but are considered deprecated and (virtually) referenced as such
in any CMake-generated Xcode project.
The following files are in place of the deprecated files:
- SuperimposeiCubHand <-> SuperimposerFactory
- SiCHCAD             <-> CADSuperimposer
- SiCHSkeleton        <-> SkeletonSuperimposer
The new files will use the recently developed SuperImpose-derived
class to draw either 3D model usign OpenGL or flat images using OpenCV.

When full and stable substituction of these files will be completed,
the deprecated files will be deleted and the SuperImpose classes
will be moved to a separate GitHub repository in order to
build a separate C++ library.