Change application and config files
There is no longer differences between using iCub_SIM or Gazebo as simulator. Thus, for the sake of simplicity, configuration files have been merged. Furthermore, there is no longer the choice of compiling the program for icubSim or the real icub. As a result, the differences lie only in the proper application config file to run.