Add splitview capability to superimpose
With this commit an overload of SICAD::superimpose() has been
implemented.

In particular, if the first input instead of being an ObjPoseMap
is a std::vector of ObjPoseMap, then each ObjPoseMap is
rendered in a separate viewport, iled horizontally with the
previous one.

Beware that to have a real splitview with the proper size for each
viewport, the OpenGL window must be initialized properly,
i.e. of the proper dimension.
Example: to render a single viewport froma camera whose size is
320x240, the OpenGL window must be 320x240. On the other hand, to have
a proper horizontal splitview of, say, three 320x240 cameras, the OpenGL
window must be initialized with a width that is three times the single
camera width, i.e. (320 * 3)x240 = 960x240.