Fix multiple viewports buf on single render
Due to a bug, if the user sets an OpenGL vindow for multiple viewports,
after infoking the single rendering version of SICAD::superimpose(),
a wrong output was given due to wrong framebuffer dimensions.

This bug is now fixed. A new SICAD::viewports_ static variable is
available and represents the number of splitviews.