Add new SICAD implementation for improving efficiency
This commit applies major refactoring to the SICAD class.
MipMap is now MIPMap. MIPmap and ObjModel typedefs are now private.
Overloaded superimpose() methods have been reimplemented focusing on reusing common code.
initOGL() is no longer a static method, this implies that any SICAD object must be initialized by the main thread.
initOGL() now enquires the GPU for the maximum size of the default framebuffer to enable using the maximum number of splitviews (by means of the scissors test) given the single image size to render.
initOGL() automatically evaluate the maximum number of splitviews adjusting it to improve efficiency.
By tiling up each of the rendered image in a big grid of images, the maximum allowed number of rendered imagesdepends on the maximum allowed texture dimension of the GPU and on the size of the single image to render. As an example, on my laptop the maximum texture dimension is 16384x16384 pixels and with 320x240 images amount to have a grid of 68x51 rendered images in parallel.