Optimize initialization of OpenGL
The test window to enquire for OpenGL for the maximum size of the renderbuffer
is now created only once, when a SICAD class is first instanziated.
To do this, a new renderbuffer_size_ static member variable has been added to
save the maximum renderbuffer size and reuse it when needed by other SICAD instances.