Fix Linux rendering bug on last opened window
There was a missing context flush while setting the projection matrix.

Context flush may be the default behavior when setting the context non-current. However this may not be the case. You can force it with a hint (may not be computationally efficient or just wrong), or you can force not to flush and issue a swapbuffer before setting the context non-current.